pyproject.toml
[project]
name = "studybattles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.134.0",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymongo[srv]>=4.16.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
    "youtube-transcript-api>=1.2.4",
]
